Implement Client Runners as Runnables and Separate Client Responsibilities
