+++ src/main/java/Application/OpenCVClient/MasterCommunicationRunner.java
@@ -3,12 +3,12 @@ import java.net.StandardSocketOptions;
 public class MasterCommunicationRunner implements Runnable {
     //TO DO: change extends Runner and instead implement Runnable
     //TO DO: reuse instances
-    private OpenCVClient activeClient;
+    private OpenCVClientOperator activeOperator;
     private ConnectionEvaluator evaluation;
     public MasterCommunicator communicator;
 
     public MasterCommunicationRunner() {
-        activeClient = OpenCVClient.getInstance();
+        activeOperator = OpenCVClientOperator.getInstance();
         evaluation = new ConnectionEvaluator();
         communicator = new MasterCommunicator();
     }
@@ -17,11 +17,11 @@ public class MasterCommunicationRunner implements Runnable {
     public void run() {
         while(true) {
             System.out.println("setup server");
-            activeClient.setup(communicator.getServer());
+            activeOperator.setup(communicator.getServer());
 
             System.out.println("start loop");
             while (evaluation.isGood()) {
-                activeClient.setServerUtilization(1);
+                activeOperator.setServerUtilization(1);
                 evaluation.evaluate();
                 try {
                     Thread.sleep(1000);
@@ -29,7 +29,7 @@ public class MasterCommunicationRunner implements Runnable {
                     e.printStackTrace();
                 }
             }
-            activeClient.setServerUtilization(0);
+            activeOperator.setServerUtilization(0);
         }
     }
 }
