+++ src/main/java/Application/OpenCVServer/ServerRunner.java
@@ -1,9 +1,11 @@
-public class ServerRunner extends Runner {
+public class ServerRunner implements Runnable {
     @Override
-    protected void runOperation() {
-        OpenCVServer activeServer = OpenCVServer.getInstance();
-        activeServer.startConnection();
-        activeServer.standbyForConnection();
-        activeServer.stopConnection();
+    public void run() {
+        while(true) {
+            OpenCVServer activeServer = OpenCVServer.getInstance();
+            activeServer.startConnection();
+            activeServer.standbyForConnection();
+            activeServer.stopConnection();
+        }
     }
 }
